Implemented open, mmap, and release file operations.
The AXI DMA device now can be opened, and can allocate a single frame to
the userspace application. For now, there is no memory allocator, so it
simply allcoates the specified size at the beginning of the contiguous
DMA region.

There is also a simple userspace test, which is essentially the port of
AXI DMA test module to userspace. Right now it confirms that the device
can mmap pages into userspace.