Added in basic support for a DMA read write transfer.
This involved adding a Read/Write IOCTL command. This command takes a
two buffers, a source and destination, and initiates a transfer across
the DMA channel, sending data from the source buffer, and receiving the
output from the PL fabric in the destination buffer.

The basic architecture support for doing this is added. Now, the scheme
used to allocate memory needs to be changed. Using `ioremap` with
`dma_map_single` will cause issues and is not scalable for this driver,
so we instead need to use the CMA allocator.