Changed the memory allocation scheme for the AXI DMA driver.
Before the AXI DMA driver expected memory to be reserved at the top of
kernel memory at boot time, via a `mem=...` kernel command line
paramter. Now, the driver exploits the contiguous memory allocater
(CMA) in the kernel to get large, contiguous memory regions for DMA
transactions. Using this method is far more robust, doesn't cause memory
to be forever reserved by the driver, and interfaces more nicely with
DMA.

Now, the kernel command line must contain a `cma=...` to tell the kernel
to reserve a region of memory for the CMA (e.g. `cma=100M` for 100 Mb).
When a application calls mmap on the AXI DMA device, it calls
`dma_alloc_coherent`, which grabs a chunk of memory from this contiguous
memory region. It then setups a VMA close callback, so when the region
is munmap'ed, it then frees the DMA region.