Completed the Read/Write IOCTL, grayscale test now successully works.
The key difficulty here was the interaction between the `mmap` and
`ioctl` calls. When the user calls mmap, the kernel returns them a
virtual address for their process (naturally). So, when they call the
ioctl, they give the kernel back a userspace virtual address.

Thus, we must go find the VMA struct for that address, and then
lookup the DMA address stored in the private data. Additionally, the
user may give us an offset from the base virtual address, so we must
then calculate this offset and add it to the DMA address.