Extended the AXI DMA module with read and write-only transactions.
The AXI DMA module now supports one way transactions, so the user can
send an IOCTL telling the DMA module to only send or receive data from
the PL fabric. This involved adding two new ioctl's, and changing the
type name for the R/W trasnaction IOCTL, so some of the examples code
also had to be changed.