Made the get DMA channel ids IOCTL more user friendly.
Before, the IOCTL required that the allocate a buffer that was at least
the size of the number of receive and transmit channels available. This
required the user to call AXIDMA_GET_NUM_DMA_CHANNELS first, so they
could allocate a buffer that size.

Now, the user can instead specify the requested number of channels, so
if they only need one channel for example, they simply need to pass in
an integer pointer. The IOCTL checks the requested number of channels;
if not enough channels are available, it returns an error to the user.