Added in probing of VDMA channels to the AXI DMA driver.
The AXI DMA driver now also searches for VDMA channels that may be
present on the PL fabric. The internal code structure of the DMA backend
of the module has been changed in prepation for adding support of VDMA
transfer. This involved mainly storing more metadata with the DMA
channel structures, and updating some IOCTL's.