Extended the driver to support an abitrary number of frame buffers.
The video write IOCTL can now support an arbitrary number of frame
buffers. So, the user now has finer-grained control over timing. This
will be particularily useful because it will allow us to chain the same
buffer back to back in the SG list, allowing the 60 Hz display to
operate at 30 Hz, matching the Kinect.