Added in generic Kinect and AXI DMA libraries.
The Kinect library serves to act as a convenient wrapper around
libfreenect, and has several convenient extensions to it. The AXI DMA
library acts as a consistent interface to the AXI DMA module's IOCTL's,
and it helps stop a lot of code repetition that was present in the
examples.

Both libraries are completely opaque, so the user need not worry about
any of the internal details.