Updated the AXI DMA transfer function to use the new utilities.
The program was also moved into the programs directory. Also, it has
been updated to handle the case when it must pass a file through a
streaming pipeline that has dependencies on additional AXI DMA channels.
In this case, it simply starts up asynchronous transactions for them,
and ignores their results.