Added in support for asynchronous notifications.
The AXI DMA driver is now able to notify userspace when an asynchronous
DMA transfer has occured. It achieves through the signal that is
specified by the user, returning the id of the channel to the userspace
`sigaction` handler. The user can set the desired signal to use through
a new IOCTL; for now, only POSIX real-time signals are permitted.

The IOCTL will set the same signal for all DMA channels, but the user
can disambguiate the source by looking at the passed in channel id (in
the field `si_int`).