Added in support for callbacks to the AXI DMA library.
As per the new feature of the AXI DMA driver, the AXI DMA library now
exposes the ability to have a callback function to the application. It
uses the signal notification from the driver to later invoke the user
context. The user context runs in a signal handling context, so they
must be careful to respect the rules for calling functions.

Also, changed the scheme for the channel ID's. The library now assumes
that the kernel driver returns a monotonic sequence of ID's starting at
zeros. Thus, it can directly use the ID's to access its internal array
structures.