Updated the AXI DMA benchmark test to deal with additional channels.
In more complex systems, there are often other DMA channels that are
along the pipeline of the endpoints you want to measure. For these, you
need to have transactions running so that they can receive the data.

In our system, if we wanted to measure the throughput of the Harris
corner detector, we would also need to startup a transaction that
receives the grayscale output from the Bayer to Grayscale converter, or
else the pipeline will hang.

Also, the benchmark test was updated to use the new libraries and
utilities.