Fixed a portability issue with the device configuration.
Certain versions of the AXI DMA driver may not implement the slave
configuration function (independent of the 3.x vs. 4.x version of the
kernel). For example, the newest Xilinx AXI DMA driver in the 4.x Linux
kernel does not implement this function. If the function is not
implemented, then it will be NULL, which will naturally cause problems.

This is solved by properly using the generic DMA layer. Rather than
directly getting the method from the DMA device structure, it is better
to call a method in the DMA API, `dmaengine_slave_config()`. This checks
that the method is not NULL, and returns `-ENOSYS` if it is. Since we
want to permit this case, we all for return codes of `ENOSYS` to be
successful.