Added in reading the device ID of channels from the Device Tree.
The code now parses the "xilnx,device-id" property of the DMA channel
device tree node entries. These are the ID's that are exposed to
userspace, and are used to uniquely identify the channel. Also, the
driver checks that all channel ID's in the nodes are unique, and
notifies the user otherwise.