Fixed a potentional issue with virtual address to PFN conversion.
The code was previous using the combination of `__phys_to_pfn()` and
`virt_to_phys()`. However, `virt_to_phys` is only meant to be used on a
kernel logical address, which is a "virtual" address in its 1-to-1
mapped virtual memory region. However, the address returned by
`dma_alloc_coherent()` returns a kernel virtual address, which exists
outside this 1-to-1 mapped region. On ARM, this happens to work, but
this is bad practice in general.