Fixed support for Linux 4.x kernels.
The 4.x kernel that was being tested against was the Analog Devices'
kernel, which had both the 3.x and 4.x version of the driver in its
repostiory. In the mainline 4.x Xilinx repo, the `struct
xilinx_dma_config` type was removed, but in Analog Devices' kernel they
kept it around. Thus, it seemed like the driver supported the 4.x
version, but was missing handling this case. This is added now, and it
supports the mainline 4.x kernel, but no longer supports Analog Devices'
kernel.