Added in the compilation of the AXI DMA library as a shared library.
The AXI DMA library now can be compiled as a shared library and
linked against with executables. Along with this, the examples were
updated to compile against the shared library, instead of needing to
directly compile with the source code.

Also, the definitions for the AXI DMA library were moved into its
Makefile, and these are now included into the other Makefiles to
give them their definitions.

This closes #8.