Enabled specification of the build options from a config Makefile.
The repository now distributes a Makefile configuration template file.
The user can specify the variables in this file, instead of on the
command line, which can quickly become cumbersome. The user has to copy
the template file to `config.mk` for the Makefile to use the variables,
which will be included into the Makefile.

The `config.mk` file is put into the gitignore, since the variables
will be specific per user. The template file `config_template.mk`
is tracked.

This closes #11.