Fixed an issue with including the `xilinx_dma.h` header file.
Between the 3.x and 4.x versions of the Xilinx kernel, the location
of the Xilinx DMA header file changed. However, in some 4.x versions
of the kernel, the header file is still in the old location from the
3.x kernel.

To deal with this issue, an option was added to the Makefile to fixup
this issue. This defines a macro from the command-line, which the
indicates to the driver to use the alternate path for the include file.

This closes #2.