Removed the start remainder transaction helpers from the code.
Earlier, the examples were augmented with code that attempted to handle
the case when the user had more complicated logic on the PL. For
example, when they have multiple input and/or output channels. However,
the code doesn't really handle these case well anyway, because the user
oftentimes requires specific amounts of data to be transfered.

In any case, this was causing confusion, because helper code internally
allocated around 35 extra MB of buffers for these "remainder"
transactions, which causes confusion when the user doesn't specify
enough CMA space on the kernel command line. Thus, this code is being
removed.