Added support for VDMA in the driver and benchmark.
The driver now allows for VDMA transfers when using the standard IOCTLs,
and the benchmark program also now supports using VDMA instead of DMA.

However, the VDMA loopback example currently isn't working, it's leading
to a hang. The issue may be with the device tree configuration, or
because the Xilinx DMA driver doesn't seem to enable frame count
interrupts.