Removed Enclustra Linux from the Travis CI file.
The Enclustra Linux repository is no longer part of the continuous
integration. At this point, the kernel is pretty old so it makes little
sense to continue supporting it. This was mainly brought about by the
fact that this kernel version doesn't have the `of_dma_configure`
function required by the last commit.