Added support for frame-based VDMA receive transfers.
This was implemented as discussed in issue #31. The software stack now
supports receive video transfers. This is a continuous receive transfer
through VDMA, or a loop transfer. Transfers like this are useful for
cameras or other video input devices that utilize frames. The call
supports an arbitrary number of frame buffers.

NOTE: This code is untested, and was only checked for correct
compilation.