Fixed a missing check for NULL in device tree parsing.
This closes #54. When parsing the child nodes of an AXI DMA node out of
the device tree, the code was missing a check for NULL, which can happen
if the node does not exist. For example, if the user mixed up the 0 and
1 index for their device tree node, this could trigger this error
condition. Without the check, the driver will crash on its probe with a
page fault.